refactor: rename extension to use `.mjs` and `.mts`
