+++ jest.config.json
@@ -8,8 +8,9 @@
       "statements": 100
     }
   },
+  "resolver": "ts-jest-resolver",
   "testMatch": ["**/*.test.ts"],
   "transform": {
-    "^.+\\.ts$": "ts-jest"
+    "^.+\\.m?ts$": ["ts-jest"]
   }
 }
